pyproject.toml
[tool.poetry]
name = "pedestrian-simulator"
version = "2025.01"
description = "Pedestrian Simulation Model for CommonRoad Scenarios"
license = "LGPL-3.0-or-later"
authors = [
    "Korbinian Moller <korbinian.moller@tum.de>",
    "Truls Nyberg <trulsny@kth.se>",
]
maintainers = [
    "Korbinian Moller <korbinian.moller@tum.de>"
]
readme = "README.md"

packages = [
    { include = "pedestrian_simulator" },
]

[tool.poetry.dependencies]

python = ">=3.10,<3.12"

commonroad-io = "2024.1"
numpy = "^1.24.2"
matplotlib = "^3.8.4"
omegaconf = "^2.3.0"
rasterio = "^1.3.10"
shapely = "^2.0.6"
tqdm = "^4.66.5"
numba = "^0.60.0"
toml= "^0.10.2"
pandas = "^2.0.3"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = ">=6.2.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"

[tool.poetry.scripts]
reactive-planner = "main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
